% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneTSP.R
\name{create_t_index_key}
\alias{create_t_index_key}
\title{Create an index mapping between the parameter vector and time steps}
\usage{
create_t_index_key(
  pars,
  data.vector.indicators,
  observation.names,
  time.step.indicators,
  cdf = FALSE
)
}
\arguments{
\item{pars}{this is the vector of parameters from the new object}

\item{data.vector.indicators}{the names of the DATA INDICATORS in the TMB template}

\item{observation.names}{the names corresponding to the observations}

\item{time.step.indicators}{the list of vectors giving the time steps to match up}
}
\description{
This function maps what time steps correspond to which parameter in the list
}
