% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneTSP.R
\name{get_yinds}
\alias{get_yinds}
\title{Get the indices corresponding to a given time step}
\usage{
get_yinds(ts, vname, t_index, grthan = FALSE)
}
\arguments{
\item{ts}{the time step to get the indices corresponding to}

\item{vname}{the vector of names to match to}

\item{t_index}{the index mapping between parameter and time step (from create_t_index_key)}

\item{grthan}{return the values greater than the time step?}
}
\description{
Get the indices corresponding to a given time step
}
