% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneTSP.R
\name{toggle_obs_gen}
\alias{toggle_obs_gen}
\title{Function to generate the function to control what variables are used during a time step}
\usage{
toggle_obs_gen(pars, t_index, data.vector.indicators, observation.names)
}
\arguments{
\item{pars}{the vector of parameters from the new object}

\item{t_index}{the index mapping between parameter and time step (from create_t_index_key)}

\item{data.vector.indicators}{the names of the DATA INDICATORS in the TMB template}

\item{observation.names}{the names corresponding to the observations}
}
\description{
Creates a function control what observations are included (non-zero weighted) during a time step as
well as set the current time step variables to a given vector of values.
}
